+++ codigo/plataforma/telemetria/libs/barometro.py
