+++ codigo/plataforma/telemetria/libs/gps.py
